backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow>=10.4.0
python-multipart==0.0.6
werkzeug==3.0.1
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
html5lib==1.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
urllib3==2.0.7 